# sstring: specialize uninitialize_string() and use resize_and_overwrite if available