- sstring: specialize uninitialized_string() using 'if constexpr'
- sstring: use resize_and_overwrite if available
- sstring: add `sstring::resize_and_overwrite()`