@tchaikov 2023-01-08
@avikivity could you help re-review this change?

- dropped the specialization
- added `sstring::resize_and_overwrite()`