@tchaikov 2023-01-08
@avikivity all comments addressed

- dropped the guard of `#if __cplusplus > 202002L`
- avoided the unnecessary `resize()` call if the new size is smaller than the original size
- added type constraints
- added `assert()` to check if the returned size is greater than specified `count`.